+++ src/main/kotlin/com/github/azurapi/azurapikotlin/internal/entities/EquipmentFit.kt
@@ -1,0 +1,7 @@
+package com.github.azurapi.azurapikotlin.internal.entities
+
+enum class EquipmentFit {
+    PRIMARY,
+    SECONDARY,
+    NONE
+}