+++ src/main/kotlin/com/github/azurapi/azurapikotlin/internal/utils/info/AtagoInfo.kt
@@ -3,6 +3,6 @@ package com.github.azurapi.azurapikotlin.internal.utils.info
 /**
  * API info
  */
-object AtagoInfo {
+internal object AtagoInfo {
     const val VERSION = "2.0.1"
 }