+++ src/main/kotlin/com/github/azurapi/azurapikotlin/internal/utils/info/TakaoInfo.kt
@@ -3,7 +3,8 @@ package com.github.azurapi.azurapikotlin.internal.utils.info
 /**
  * JSON deserializer info
  */
-object TakaoInfo {
-    const val JSON_SOURCE = "https://raw.githubusercontent.com/AzurAPI/azurapi-js-setup/master/ships.json"
+internal object TakaoInfo {
+    const val JSON_SHIP = "https://raw.githubusercontent.com/AzurAPI/azurapi-js-setup/master/ships.json"
+    const val JSON_EQUIPMENT = "https://raw.githubusercontent.com/AzurAPI/azurapi-js-setup/master/equipments.json"
     const val JSON_VERSION = "https://raw.githubusercontent.com/AzurAPI/azurapi-js-setup/master/version-info.json"
 }